idxd: In perf tool, correctly pass fill pattern as a uint64_t
The pattern is 64 bits but we were only passing in 8.

Fixes #2821

Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/17686
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot

+++ examples/idxd/perf/perf.c
@@ -523,7 +523,7 @@ _submit_single(struct idxd_chan_entry *t, struct idxd_task *task)
 		/* For fill use the first byte of the task->dst buffer */
 		diov.iov_base = task->dst;
 		diov.iov_len = g_xfer_size_bytes;
-		rc = spdk_idxd_submit_fill(t->ch, &diov, 1, *(uint8_t *)task->src,
+		rc = spdk_idxd_submit_fill(t->ch, &diov, 1, *(uint64_t *)task->src,
 					   flags, idxd_done, task);
 		break;
 	case IDXD_CRC32C:
